I create child component for token section.
